+++ src/main/controllers/features/SoleTrader/nameController.ts
