+++ src/main/controllers/index.ts
@@ -1,17 +1,19 @@
+// Sole trader
 export * as indexController from "./features/SoleTrader/indexController";
-export * as soleTraderDateOfBirthController from "./features/SoleTrader/soleTraderDateOfBirthController";
-export * as soleTraderNameController from "./features/SoleTrader/soleTraderNameController";
-export * as soleTraderWhatIsYourRoleController from "./features/SoleTrader/soleTraderWhatIsYourRoleController";
-export * as soleTraderWhereDoYouLiveController from "./features/SoleTrader/soleTraderWhereDoYouLiveController";
-export * as soleTraderCorrespondenceAddressManualController from "./features/SoleTrader/soleTraderCorrespondenceAddressManualController";
-export * as soleTraderCorrespondenceAddressAutoLookupController from "./features/SoleTrader/soleTraderCorrespondenceAddressAutoLookupController";
-export * as soleTraderSectorYouWorkInController from "./features/SoleTrader/soleTraderSectorYouWorkInController";
-export * as whichsectorOtherController from "./features/SoleTrader/WhichSectorOtherController";
-export * as soleTraderCorrespodanceAddressDetailsController from "./features/SoleTrader/soleTraderCorrespodanceAddressDetailsController";
-export * as soleTraderCorrespondenceAddressConfirmController from "./features/SoleTrader/soleTradercorrespondenceAddressConfirmController";
-export * as soleTraderNationalityController from "./features/SoleTrader/soleTraderNationalityController";
+export * as soleTraderDateOfBirthController from "./features/SoleTrader/dateOfBirthController";
+export * as soleTraderNameController from "./features/SoleTrader/nameController";
+export * as soleTraderWhatIsYourRoleController from "./features/SoleTrader/whatIsYourRoleController";
+export * as soleTraderWhereDoYouLiveController from "./features/SoleTrader/whereDoYouLiveController";
+export * as soleTraderCorrespondenceAddressManualController from "./features/SoleTrader/correspondenceAddressManualController";
+export * as soleTraderCorrespondenceAddressAutoLookupController from "./features/SoleTrader/correspondenceAddressAutoLookupController";
+export * as soleTraderSectorYouWorkInController from "./features/SoleTrader/sectorYouWorkInController";
+export * as whichSectorOtherController from "./features/SoleTrader/whichSectorOtherController";
+export * as soleTraderCorrespodanceAddressDetailsController from "./features/SoleTrader/correspodanceAddressDetailsController";
+export * as soleTraderCorrespondenceAddressConfirmController from "./features/SoleTrader/correspondenceAddressConfirmController";
+export * as soleTraderNationalityController from "./features/SoleTrader/nationalityController";
 export * as typeOfBusinessController from "./features/SoleTrader/typeOfBusinessController";
 export * as OtherTypeOfBusinessController from "./features/SoleTrader/otherTypeOfBusinessController";
+
 export * as healthCheckController from "./healthCheckController";
 export * as companyLookupController from "./features/limited/companyLookupController";
 export * as nameRegisteredWithAmlController from "./features/limited/nameRegisteredWithAmlController";
@@ -19,7 +21,7 @@ export * as businessMustbeAmlRegisteredController from "./features/limited/busin
 export * as companyInactiveController from "./features/limited/companyInactiveController";
 export * as whatIsTheBusinessNameController from "./features/unincorporated/whatIsTheBusinessNameController";
 export * as accessibilityStatementController from "./accessibilityStatementController";
-export * as soleTraderWhatIsTheBusinessNameController from "./features/SoleTrader/soleTraderWhatIsTheBusinessNameController";
+export * as soleTraderWhatIsTheBusinessNameController from "./features/SoleTrader/whatIsTheBusinessNameController";
 export * as unincorporatedWhatIsYourNameController from "./features/unincorporated/unincorporatedWhatIsYourNameController";
 export * as unincorporatedNameRegisteredWithAmlController from "./features/unincorporated/unincorporatedNameRegisteredWithAmlController";
 export * as unincorporatedWhatIsYourRoleController from "./features/unincorporated/unincorporatedWhatIsYourRoleController";
